refactor(editor): make schema subpaths package-owned
